Simplify/Improve FlatArray design and impl
